+++ app/components/ClientOnly.tsx
@@ -2,7 +2,6 @@
 
 import React, { useState, useEffect } from "react";
 import Loader from "./Loader";
-import "@/app/i18n";
 
 interface ClientOnlyProps {
   children: React.ReactNode;