+++ app/hooks/useGlobalErrorHandler.ts
@@ -1,6 +1,5 @@
 "use client";
 
-import { useRouter } from "next/navigation";
 import { useCallback } from "react";
 import toast from "react-hot-toast";
 import { useTranslation } from "react-i18next";
@@ -8,7 +7,6 @@ import { useAuth } from "../context/AuthContext";
 import useSignOut from "./useSignOut";
 
 const useGlobalErrorHandler = () => {
-  const router = useRouter();
   const { authStatus } = useAuth();
   const { signOut } = useSignOut();
   const { t } = useTranslation();
@@ -18,21 +16,21 @@ const useGlobalErrorHandler = () => {
       if (error) {
         if (error?.response?.status === 401) {
           if (authStatus) {
-            await signOut();
+            signOut();
           }
         } else {
-          const errorCode: string = error?.response?.data?.code ?? "00000";
-          console.log(
-            "useGlobalErrorHandler: " + JSON.stringify(error, null, 2)
-          );
+          const errorMessage: string =
+            error?.response?.data?.code ??
+            error?.response?.data?.message ??
+            error?.message;
 
-          toast.error(t(errorCode));
+          toast.error(t(errorMessage ?? "Something_went_wrong"));
         }
       } else {
-        toast.error(t("00000"));
+        toast.error(t("Something_went_wrong"));
       }
     },
-    [router]
+    [authStatus, signOut, t]
   );
 
   return { handleError };