+++ app/layout.tsx
@@ -1,15 +1,13 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+import { Montserrat } from "next/font/google";
 import "./globals.css";
-import ToasterProvider from "./providers/ToasterProvider";
-import ClientOnly from "@/app/components/ClientOnly";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import TanStackReactQueryProvider from "./providers/TanStackReactQueryProvider";
-import { AuthProvider } from "./context/AuthContext";
-import { GoogleOAuthProvider } from "@react-oauth/google";
+import AppClient from "./AppClient";
 
-const inter = Inter({ subsets: ["latin"] });
+const montserrat = Montserrat({
+  subsets: ["latin", "cyrillic"],
+});
 
 export const metadata: Metadata = {
   title: "PawPal",
@@ -23,17 +21,8 @@ export default async function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>
-        <AuthProvider>
-          <TanStackReactQueryProvider>
-            <GoogleOAuthProvider clientId="508483481007-bnvs2k6jpc0ei0t075ecps22b5cbga8f.apps.googleusercontent.com">
-              <ClientOnly>
-                <ToasterProvider />
-                <div>{children}</div>
-              </ClientOnly>
-            </GoogleOAuthProvider>
-          </TanStackReactQueryProvider>
-        </AuthProvider>
+      <body className={montserrat.className}>
+        <AppClient>{children}</AppClient>
         <Analytics />
         <SpeedInsights />
       </body>